KA-160 Context specificity renaming
